restructured project folder to focus on vs code extension, moved old files into D3 folder
