+++ D3/webpack.config.js
@@ -1,0 +1,27 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  mode: "development",
+  entry: path.resolve(__dirname, "./src/index.js"),
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'bundle.js'
+  },
+  devServer: {
+    static: {
+      directory: path.resolve(__dirname, 'src')
+    },
+    open: true,
+    hot: true,
+    port: 8080,
+  },
+  resolve: {
+    extensions: [".js", ".json"],
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, "src", "index.html")
+    })
+  ]
+}
