+++ src/extension.ts
@@ -16,7 +16,7 @@ export function activate(context: vscode.ExtensionContext) {
 	const disposable = vscode.commands.registerCommand('nextFlow.helloWorld', () => {
 		// The code you place here will be executed every time your command is executed
 		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from nextFlow!');
+		vscode.window.showInformationMessage('Hello VS Code!');
 	});
 
 	context.subscriptions.push(disposable);
