+++ vsc-extension-quickstart.md
